[refactor] entities → domain 패키지 분리
domain에 비즈니스 로직을 구현하여 미래에 웹 구현에서 핵심 로직 재사용성을 높이기 위한 분리.

- entities 패키지 → domain 패키지로 분리 및 파일 이동
- Auth/User/Mock 타입 파일 domain/src로 리네임 및 index.ts 추가
- 기존 entities/src/index.ts, tsconfig.json 등 삭제/이동
- auth 서비스: Apple/GoogleAuthService 파일명 통일 및 리팩터링
- Zustand 인증 스토어, 훅, localStorage 등 구조 개선
- API/유저/Mock 관련 함수 및 테스트 코드 경로/구조 정리
- Storybook, Metro, Babel, tsconfig 등 경로/alias 업데이트
