+++ apps/mobile/package.json
@@ -26,7 +26,7 @@
     "@gorhom/bottom-sheet": "5.2.7",
     "@mockly/api": "workspace:*",
     "@mockly/design-system": "workspace:*",
-    "@mockly/entities": "workspace:*",
+    "@mockly/domain": "workspace:*",
     "@mockly/typescript-config": "workspace:*",
     "@mockly/utils": "workspace:*",
     "@react-native-async-storage/async-storage": "^2.2.0",
