+++ apps/mobile/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@mockly/entities": ["../../packages/entities/src"],
+      "@mockly/domain": ["../../packages/domain/src"],
       "@mockly/design-system": ["../../packages/design-system/src"],
       "@mockly/api": ["../../packages/api/src"],
       "@mockly/config": ["../../packages/config/src"],
