+++ packages/api/package.json
@@ -28,7 +28,7 @@
     "typescript": "5.9.3"
   },
   "dependencies": {
-    "@mockly/entities": "workspace:*",
+    "@mockly/domain": "workspace:*",
     "axios": "1.13.2"
   }
 }
