+++ packages/api/src/auth/loginWithPKCECode.ts
@@ -1,4 +1,4 @@
-import { AuthToken, LocationInfo, DeviceInfo, AuthProvider } from '@mockly/entities';
+import { AuthToken, LocationInfo, DeviceInfo, AuthProvider } from '@mockly/domain';
 import { apiClient } from '../client';
 
 interface GoogleLoginReqDTO {
