+++ packages/api/src/auth/renewalToken.ts
@@ -1,4 +1,4 @@
-import { AccessRefreshToken, LocationInfo } from '@mockly/entities';
+import { AccessRefreshToken, LocationInfo } from '@mockly/domain';
 import { apiClient } from '../client';
 interface RenewalTokenResDTO {
   accessToken: string;
