+++ packages/api/src/mock/getMocks.ts
@@ -1,4 +1,4 @@
-import { Mock } from '@mockly/entities';
+import { Mock } from '@mockly/domain';
 import { apiClient } from '../client';
 
 interface MockResDTO {
