+++ packages/domain/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mockly/entities",
+  "name": "@mockly/domain",
   "version": "0.0.1",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
