+++ packages/domain/src/auth/auth.type.ts
