+++ packages/domain/src/mock/index.ts
@@ -1,0 +1,1 @@
+export * from './mock.type';
