+++ packages/domain/src/mock/mock.type.ts
