+++ packages/domain/src/user/index.ts
@@ -1,0 +1,1 @@
+export * from './user.type';
