--- packages/entities/src/index.ts
@@ -1,4 +1,0 @@
-// Export all entities
-export * from './User';
-export * from './Mock';
-export * from './Auth';
